refactor: resource form body creation
